[r=andreserl][bug=][author=smoser] update packaging to pass --ip to maas configure_master_dhcp
This makes a change to use debconf to get the (now required) ip address of the dhcp server, via the debconf name 'maas-dhcp/maas-dhcp-addr'. If left blank, it will
pick the ip address of the named interface from the provided maas-dhcp-interface.

There are some other changes lumped in here also.
 * fix a lintian warning in maas-dhcp.postrm
 * move tests/maas-integration.py to a patch in debian/patches/
   Debian packages are not supposed to modify the trunk directly, and
   'debuild -S' complained about changes to that file that were not accounted for.