[r=smoser][bug=][author=jtv] Upstart script for custom instance of dhcpd, using custom config, pidfile, leases file, and interfaces config so that we can dictate which interfaces it should listen on.
The interfaces config is to be written into /var/lib/maas/dhcpd-interfaces by the write_dhcp_config task.  (It's not quite static config, so it belongs in /var/lib rather than /etc).  Until it exists, it would be foolish or even harmful to start dhcpd; in that case the script aborts.  The write_dhcp_config task will restart it when (re)configuring dhcpd.